+++ app/App.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import {NavigationContainer} from "@react-navigation/native";
 import LoginNavigator from "./components/navigator/LoginNavigator";
-import {PROFILE_EMPTY, TOKEN_EMPTY} from "./utils/Login/InitialScreenUtils";
+import {PROFILE_EMPTY, TOKEN_EMPTY} from "./utils/Login/LoginUtils";
 import LoginContext from "./context/LoginContext";
