+++ app/utils/Login/InitialScreenUtils.ts
@@ -1,6 +1,3 @@
-export const TOKEN_EMPTY : string = "token has not fetched";
-export const PROFILE_EMPTY : {[key:string] : string} = {
-	id: "profile has not fetched",
-	email: "profile has not fetched",
-	profile_image_url: "",
-};
+const LOGIN_BUTTON_TEXT : string = "카카오톡으로 로그인";
+
+export default LOGIN_BUTTON_TEXT;
