+++ app/utils/Login/LoginUtils.ts
@@ -1,0 +1,6 @@
+export const TOKEN_EMPTY : string = "token has not fetched";
+export const PROFILE_EMPTY : {[key:string] : string} = {
+	id: "profile has not fetched",
+	email: "profile has not fetched",
+	profile_image_url: "",
+};
