--- app/utils/utilsTest.js
@@ -1,1 +1,0 @@
-console.log("util Test");
